Generate unique artifacts for Xray data generation
At the beginning of the test suite we create several artifacts locally rather than using the same artifacts many times. We delete these local artifacts at the end of the test.

artifact.zip and the gitignore are updated. This ensures that other tests still have access to artifact.zip and that this commit does not break other parts of the test.